api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "The API for the Panso.se project"
authors = ["Joakim Hellsén <tlovinator@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
loguru = "^0.7.2"
python-dotenv = "^1.0.0"
apscheduler = "^3.10.4"
django = "^4.2.7"
django-ninja = "^1.0.1"
django-simple-history = "^3.4.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
pytest = "^7.4.3"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = [
  "poetry-core",
]

[tool.ruff]
exclude = ["migrations"]
fix = true
unsafe-fixes = true
preview = true
select = ["ALL"]
ignore = [
  "D100",
  "D104",
  "CPY001",
  "ERA001",
  "TD002",
  "TD003",
  "FIX002",
  "COM812",
  "ISC001",
]
line-length = 120

[tool.ruff.pydocstyle]
convention = "google"

[tool.ruff.per-file-ignores]
"**/tests/**/*.py" = [
  "S101", # asserts allowed in tests...
  "ARG",  # Unused function args -> fixtures nevertheless are functionally relevant...
  "FBT",  # Don't care about booleans as positional arguments in tests, e.g. via @pytest.mark.parametrize()
]
